SnapdragonCamera: Fix OutOfBoundsException in monkey test
In monkey test, set defaultValue operation in xml file, parsing
XML in framework code, pref.getValue() return null when
defaultValue not ready. When findIndexOfValue() failed to get
the index of the value, it will return -1, but -1 can`t be the
index of the array, it will cause exception. So if can`t find
the index of the value, just go back and don`t execute the
following code.

Change-Id: I2b5595d7d558e3ab3863287e3d1424df141c5d59
CRs-Fixed: 1063228
Signed-off-by: DarkKnight6499 <yazad.madan@gmail.com>