SnapdragonCamera: Fix Camera1 force close when refocusing the picture.
When the screen is PORTRAIT mode, we get the screen always height < width.
But we get the image always height < width.Under this circumstance,
the inSampleSize maybe incorrect , which will cause OOM when create bitmap.

Need to check the screen's orientation,and choose the correct width and
height to calculate the inSampleSize.

Change-Id: I11743c862ebdb9d35c4fc5de7bf40003c69b7219
CRs-Fixed: 2013250
Signed-off-by: DarkKnight6499 <yazad.madan@gmail.com>